+++ SolsUnderground/Enemy.cs
@@ -68,7 +68,7 @@ namespace SolsUnderground
 
         //method to be overridden
         //changes health when hit by the player
-        public abstract void TakeDamage(int damage, double knockback);
+        public abstract void TakeDamage(int damage, int knockback);
 
         public abstract void EnemyMove(Player player);
     }
